Complete separation of direct X add more function to transform fix the way star create work

+++ SpaghettiEngine/Framework/Utilities/CMessageBox.cpp
@@ -36,7 +36,7 @@ std::string CMessageBox::GetTypeStr() const noexcept
 	return StringConverter::WStrToStr(_type);
 }
 
-float CMessageBox::GetFlag() const noexcept
+long CMessageBox::GetFlag() const noexcept
 {
 	return _flag;
 }
@@ -61,7 +61,7 @@ void CMessageBox::SetType(const std::string& type) noexcept
 	_type = StringConverter::StrToWStr(type);
 }
 
-void CMessageBox::SetFlag(const float& flag) noexcept
+void CMessageBox::SetFlag(const long& flag) noexcept
 {
 	_flag = flag;
 }
